feat: refactor Model3D nodes for improved structure and functionality - Introduced a comprehensive refactor plan for the `model3d` nodes, addressing real bugs and design smells. - Created a new abstract base class `GlbTransformNode` to streamline node implementations and eliminate redundancy. - Split the monolithic `model3d` file into multiple focused files for better maintainability. - Added new nodes for loading, saving, and converting 3D models, enhancing user interaction. - Implemented safety net tests and fixed existing bugs in metadata handling and model processing. - Expanded tests to validate the functionality of new and refactored nodes, ensuring robust performance.
